Fix a bug in re-positioning dragged when leaving the list from the top  (#68)
* Add utils file to get insrances

* Add new tests cases with a new description

* Split different cases into multiple files

* Fix a bug in updating isMovingDownPrev & remove unnecessary comparison when the flag is zero

* Fix a bug in isDraggedHasNoEffect

* Add a more efficient way to manage dragged motions

* Add enhancements fixing leaving from top issues

* Remove old test cases
